+++ src/test/kotlin/kotools/types/number/StrictlyPositiveIntTest.kt
@@ -151,9 +151,7 @@ class StrictlyPositiveIntTest {
             // GIVEN
             val x: StrictlyPositiveInt = StrictlyPositiveInt.range.random()
                 .toStrictlyPositiveInt()
-            val y: NonZeroInt = NonZeroInt.ranges.random()
-                .random()
-                .toNonZeroInt()
+            val y: NonZeroInt = NonZeroInt.random
             // WHEN
             val result: NonZeroInt = x * y
             // THEN